% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flicker-package.R
\docType{package}
\name{flicker}
\alias{flicker}
\alias{flicker-package}
\title{flicker: Common Wrappers and Missing Functionality from 'sparklyr' and 'dplyr'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

'flicker' adds convenience wrappers for common 'dplyr' pipelines but also ensures they will work within a
    Spark context.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nathaneastwood/flicker}
  \item Report bugs at \url{https://github.com/nathaneastwood/flicker/issues}
}

}
\author{
\strong{Maintainer}: Nathan Eastwood \email{nathan.eastwood@icloud.com}

}
\keyword{internal}
