% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.tbl_lazy}
\alias{is.tbl_lazy}
\alias{is.tbl_spark}
\title{Lazy R Object Checks}
\usage{
is.tbl_lazy(x)

is.tbl_spark(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A \code{logical(1)}.
}
\description{
Lazy R Object Checks
}
