% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na.R
\name{remove_na}
\alias{remove_na}
\title{Remove NAs}
\usage{
remove_na(.data, ..., .verbose = FALSE)
}
\arguments{
\item{.data}{A \code{tbl_spark} or a \code{data.frame}.}

\item{...}{\code{symbols}. The column(s) from which \code{NA}s should be removed. If no columns are provided, all \code{NA}s will be
removed.}

\item{.verbose}{\code{logical(1)}. If \code{TRUE}, return a message recording the number of rows removed and the number of
rows remaining (default: \code{FALSE}).}
}
\value{
An object of the same type as \code{.data}. The output has the following properties:
\itemize{
\item Rows are a subset of the input, but appear in the same order.
\item Columns are not modified.
\item The number of groups will be the same.
\item Attributes are preserved.
}
}
\description{
A helper function to remove any rows where \code{NA} values appear in the specified column(s).
}
\examples{
# You can remove NAs from a single column
remove_na(airquality, Ozone)

# Or from multiple columns
remove_na(airquality, Ozone, Solar.R)

# By default, the function will remove any rows where NA values
# appear in any column
remove_na(airquality)

}
