% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_select.R
\name{union_select}
\alias{union_select}
\title{Union Select Columns From Multiple Datasets}
\usage{
union_select(data, cols = NULL, all = TRUE)
}
\arguments{
\item{data}{A \code{list()} of \code{data.frame}s or Spark \code{DataFrame}s.}

\item{cols}{\code{character(n)}. The name(s) of the columns to return (def: \code{NULL}).}

\item{all}{\code{logical(1)}. Whether to keep duplicate records (def: \code{TRUE}) or not (\code{FALSE}).}
}
\value{
A Spark \code{DataFrame} or a \code{data.frame} depending on the input, \code{.data}.
}
\description{
This function will union the records from multiple data sets returning only the requested columns (all of which are
assumed to be named the same between data sets).
}
\examples{
a <- data.frame(col1 = c(1:10, 10), col2 = 6)
b <- data.frame(col1 = c(1:5, 5), col2 = 4)
c <- data.frame(col1 = c(0, 1, 1, 2, 3, 5, 8))

# You can union specific columns
union_select(data = list(a, b, c), cols = "col1")

# And you can remove duplicate records
union_select(data = list(a, b, c), cols = "col1", all = FALSE)

}
