% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkplugs-package.R
\docType{package}
\name{sparkplugs}
\alias{sparkplugs}
\alias{sparkplugs-package}
\title{sparkplugs: Common Wrappers and Missing Functionality from 'sparklyr' and 'dplyr'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Plugging the gaps of 'sparklyr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nathaneastwood/sparkplugs}
  \item Report bugs at \url{https://github.com/nathaneastwood/sparkplugs/issues}
}

}
\author{
\strong{Maintainer}: Nathan Eastwood \email{nathan.eastwood@icloud.com}

}
\keyword{internal}
